+++ JisooX/modules/sql/feds_sql.py
@@ -1,6 +1,6 @@
 import threading
 
-from sqlalchemy import Column, String, UnicodeText, func, distinct, Integer, Boolean
+from sqlalchemy import Column, String, UnicodeText, func, distinct, Integer, Boolean, BigInteger
 from telegram.error import BadRequest, TelegramError, Unauthorized
 
 from JisooX import dispatcher
@@ -56,7 +56,7 @@ class BansF(BASE):
 
 class FedsUserSettings(BASE):
 	__tablename__ = "feds_settings"
-	user_id = Column(Integer, primary_key=True)
+	user_id = Column(BigInteger, primary_key=True)
 	should_report = Column(Boolean, default=True)
 
 	def __init__(self, user_id):
